Rename client/builder to client/core (ref #721)
