+++ client/core/core.go
@@ -1,4 +1,4 @@
-package builder
+package core
 
 import (
 	"fmt"
