+++ x/stake/commands/query.go
@@ -11,7 +11,7 @@ import (
 
 	crypto "github.com/tendermint/go-crypto"
 
-	"github.com/cosmos/cosmos-sdk/client/builder"
+	"github.com/cosmos/cosmos-sdk/client/core"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire" // XXX fix
 	"github.com/cosmos/cosmos-sdk/x/stake"
@@ -47,7 +47,7 @@ func GetCmdQueryCandidates(cdc *wire.Codec, storeName string) *cobra.Command {
 
 			key := PrefixedKey(stake.MsgType, stake.CandidatesKey)
 
-			res, err := builder.Query(key, storeName)
+			res, err := core.Query(key, storeName)
 			if err != nil {
 				return err
 			}
@@ -87,7 +87,7 @@ func GetCmdQueryCandidate(cdc *wire.Codec, storeName string) *cobra.Command {
 
 			key := PrefixedKey(stake.MsgType, stake.GetCandidateKey(addr))
 
-			res, err := builder.Query(key, storeName)
+			res, err := core.Query(key, storeName)
 			if err != nil {
 				return err
 			}
@@ -133,7 +133,7 @@ func GetCmdQueryDelegatorBond(cdc *wire.Codec, storeName string) *cobra.Command 
 
 			key := PrefixedKey(stake.MsgType, stake.GetDelegatorBondKey(delegator, addr, cdc))
 
-			res, err := builder.Query(key, storeName)
+			res, err := core.Query(key, storeName)
 			if err != nil {
 				return err
 			}
@@ -175,7 +175,7 @@ func GetCmdQueryDelegatorBonds(cdc *wire.Codec, storeName string) *cobra.Command
 
 			key := PrefixedKey(stake.MsgType, stake.GetDelegatorBondsKey(delegator, cdc))
 
-			res, err := builder.Query(key, storeName)
+			res, err := core.Query(key, storeName)
 			if err != nil {
 				return err
 			}
